+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Change log
 - Refactored parameter definition support.
 - Updated `Krizalys\Onedrive\Proxy\FileProxy`.
 - Updated `Krizalys\Onedrive\Proxy\FolderProxy`.
+- Updated `Krizalys\Onedrive\Proxy\FileSystemInfoProxy`.
 
 [2.4.2] - 2019-09-16
 --------------------